Fix register after multiple attempts with invalid email. (#63)
Due to the way the underlying API works, we can get into a situation where the identify call succeeds, and we have a proceed context that's in the wrong state.

I originally used the begin context from the login page to optimize network calls. But it causes issues due to the way the API/SDK is implemented.

See OKTA-403240