Listen for config changes in Foreground Activity.
This fixes a bug in Android 12 (API 31) which occurs when the host Activity (MainActivity) has a fixed orientation (landscape), but the Chrome Custom Tabs opens in portrait.

Since we don't have anything related to these config changes in our activities, it's safe to not recreate our activity when these happen.